Improve o-tooltip support in the subnav (#263)
Improve o-tooltip support in the subnav by enabling visible overflow.

It works on the basis that the absolutely positioned tooltip will be relative to its closest positioned ancestor. By removing `position: relative` from `.o-header__subnav-list+.o-header__subnav-list` the subnav tooltip will be positioned relative to `.o-header__container`, a parent of `o-header__subnav-wrap-outside` which has `overflow:hidden`.

The `position: relative` we remove from `.o-header__subnav-list+.o-header__subnav-list` is used to achieve the subnav break. As that no longer works we have to use the first child of the subnav and some awkward absolute values to achieve the same break height.