Make the height of the sticky nav public.
Making the sticky nav height public will allow users to position
elements below the sticky nav, in cases where the element must
align with other elements in the page and thus cannot be a child
of the sticky `o-header` element.

This is a pretty niche usecase, but better than projects using a
hard coded value!